requirements.txt
blinker==1.4
click==7.1.2
dataclasses-json==0.5.4
dnspython==2.1.0
email-validator==1.1.2
et-xmlfile==1.0.1
Flask==1.1.2
Flask-Cors==3.0.10
Flask-Login==0.5.0
Flask-Principal==0.4.0
Flask-Security-Too==4.0.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.14.3
greenlet==1.0.0
idna==3.1
itsdangerous==1.1.0
jdcal==1.4.1
Jinja2==2.11.3
MarkupSafe==1.1.1
marshmallow==3.11.1
marshmallow-enum==1.5.1
mypy-extensions==0.4.3
ngram==3.3.2
openpyxl==3.0.5
passlib==1.7.4
psycopg2-binary==2.8.6
PyJWT==2.0.1
PyYAML==5.4.1
six==1.15.0
SQLAlchemy==1.4.4
stringcase==1.2.0
typing-extensions==3.7.4.3
typing-inspect==0.6.0
-e git+https://github.com/robertradnai/vocabulary-lib@58c77b06c81d93ee5c512fe7ac910dd45e091031#egg=vocabulary_RR
Werkzeug==1.0.1
WTForms==2.3.3
